move objectvisibletask
